requirements.txt
CherryPy==5.4.0
PyNaCl==1.0.1
PyYAML==3.11
SQLAlchemy==1.0.12
cbor==1.0.0
cffi==1.6.0
click==6.6
coverage==4.0.3
docopt==0.6.2
parse==1.6.6
psycopg2==2.6.1
py==1.4.31
pycparser==2.14
pytest==2.9.1
pytest-cov==2.2.1
records==0.4.3
six==1.10.0
tablib==0.11.2
